pyproject.toml
[project]
name = "narad"
version = "0.1.0"
description = "Real-time geopolitical news aggregation dashboard"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "httpx>=0.28",
    "feedparser>=6.0",
    "pydantic-settings>=2.5",
    "rapidfuzz>=3.10",
    "apscheduler>=3.10,<4.0",
    "python-dateutil>=2.9",
    "scikit-learn>=1.4",
    "google-genai>=1.0",
]

[tool.setuptools.packages.find]
include = ["narad*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]
